refact: refactored the disjointed modules into a library
Modified the stochastic_modeling.py module to be more modulirized. For
this added the utils.py complementary module. Also added files required
for this to be a library (setup.py), moved the modules into a new
directory, which contains the library.

Also, initiated the refactoring according to the guidelines stablished
by GIGC.